server/pyproject.toml
[tool.poetry] # set for API apps (default is aimed at distrubuted packages)
package-mode = false

[project]
name = "python_server"
version = "0.1.0"
description = "A basic python server"
authors = [{ name = "victor loginov" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["flask (>=3.1.1,<4.0.0)", "python-dotenv (>=1.1.1,<2.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
